khkw/No40-log-example-python/requirements.txt
kafka-python==2.0.1
